Update UDF
This source code contains the update for use the UDF. Using the UDF.c
file you can make your own custom function and call it like any other
function of AccessB.

You must write your own function inside UDF function, and your ISR
function inside of UDF_ISR and then compile the code in MPLAB, the hex
file can be programmed with UDF programmer that is in the UDF programmer
repo.

You must consider read all the comments in the UDF.c file there is a
special note about size of code and the ISR priority that must be set to
LOW PRIORITY, this because high priority is for USB module only.